ci: use Debian instead of deprecated i386/ubuntu
Ubuntu has ended support for 32 bit platforms and is not maintaining any
release anymore that has 32 bit support. But we still use i386/ubuntu in
our CI pipeline to test for compatibility with 32 bit systems, even
though that specific image does not receive updates anymore.

Besides being end-of-life, this image also doesn't have all packages
available to it anymore. This creates problems with a subsequent patch,
where we're about to pull in cbindgen for generating Rust to C bindings.

Drop the Ubuntu image and use Debian instead, which continues to
maintain its 32 bit port.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>